+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
         - Link documents to a resident and save them withing the app data
         - Display documents in the document tab of the resident information
         - Include generated contracts to the managed documents
+    - Functionality to increase rent 
 - Contract
     - Display resident email
     - Additional agreements section